TRI-140 validator 리팩토링
