+++ src/main/java/com/trinity/ctc/kakao/service/KakaoApiService.java
@@ -8,12 +8,7 @@ import com.trinity.ctc.util.exception.CustomException;
 import com.trinity.ctc.util.exception.error_code.KakaoErrorCode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
+import org.springframework.http.*;
 import org.springframework.stereotype.Service;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
