+++ src/main/java/com/trinity/ctc/util/formatter/JsonUtil.java
@@ -3,8 +3,8 @@ package com.trinity.ctc.util.formatter;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.trinity.ctc.util.exception.CustomException;
 import com.trinity.ctc.util.exception.error_code.JsonParseErrorCode;
+import com.trinity.ctc.util.exception.CustomException;
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 import lombok.RequiredArgsConstructor;
